requirements.txt
numpy~=1.23.5
scipy~=1.10.1
tensorflow-gpu~=2.6.0
scikit-learn~=1.2.2
matplotlib~=3.4.3
setproctitle~=1.3.2